refact!: move common logic to CommonCLI
